+++ src/tools/sections/get_sections.ts
@@ -1,8 +1,8 @@
 import { z } from "zod";
 import fs from "fs";
-import { TestRailClient } from "../client/testrail.js";
-import { ToolDefinition } from "../types/custom.js";
-import { SectionSchema } from "../types/testrail.js";
+import { TestRailClient } from "../../client/testrail.js";
+import { ToolDefinition } from "../../types/custom.js";
+import { SectionSchema } from "../../types/testrail.js";
 
 const parameters = {
     project_id: z.number().describe("The ID of the project. Use get_projects to find available projects"),